Allow the performing the SET command if the first entry is a parameter plus and equals sign and a value. It will still accept it if the SET command precedes it as well.

+++ src/main/java/com/mycompany/amazonlogger/LoopParam.java
@@ -26,7 +26,7 @@ public class LoopParam {
         try {
             Integer iVal = Utils.getHexValue(name);
             if (iVal == null) {
-                iVal = Utils.getIntValue(name);
+                iVal = Utils.getIntValue(name).intValue();
             }
             this.value = iVal;
             this.paramName = null;
